Support no-disk uploads
Workaround when the image to be uploaded is only in memory and
not on disk (e.g. read only filesystems like Heroku). Instead of
a bare string as the filename, pass in an object which returns
the filename via to_s as expected, and responds to the
preloaded_image_data method with the binary representation.